Rename `ty/generate` module as `tr_impl`.
