+++ modules/rust/gen/src/tr_impl.rs
@@ -1,0 +1,3 @@
+mod class;
+mod json_ld;
+mod rdf;
