+++ modules/rust/gen/src/tr_impl/class/enum.rs
