+++ modules/rust/gen/src/tr_impl/class/primitive.rs
