+++ modules/rust/gen/src/tr_impl/rdf/from.rs
