+++ modules/rust/gen/src/tr_impl/rdf/from/enum.rs
