+++ modules/rust/gen/src/tr_impl/rdf/from/struct.rs
