+++ modules/rust/gen/src/tr_impl/rdf/to/struct.rs
