+++ modules/rust/gen/src/ty/alias.rs
@@ -1,5 +1,9 @@
 use proc_macro2::Ident;
-use treeldr::TId;
+use quote::format_ident;
+use rdf_types::Vocabulary;
+use treeldr::{BlankIdIndex, IriIndex, TId};
+
+use crate::{syntax, Context, Error, GenerateSyntax, Scope};
 
 use super::Parameters;
 
@@ -46,5 +50,39 @@ impl Alias {
 		mut dependency_params: impl FnMut(TId<treeldr::Layout>) -> Parameters,
 	) -> Parameters {
 		dependency_params(self.target)
+	}
+}
+
+impl<M> GenerateSyntax<M> for Alias {
+	type Output = syntax::Alias;
+
+	fn generate_syntax<V: Vocabulary<Iri = IriIndex, BlankId = BlankIdIndex>>(
+		&self,
+		context: &Context<V, M>,
+		scope: &Scope,
+	) -> Result<Self::Output, Error> {
+		let params = syntax::LayoutParameters {
+			identifier: if self.params().identifier {
+				Some(format_ident!("I"))
+			} else {
+				None
+			},
+		};
+
+		let mut scope = scope.clone();
+		scope.params.identifier = params.identifier.clone().map(|i| {
+			syn::Type::Path(syn::TypePath {
+				qself: None,
+				path: i.into(),
+			})
+		});
+
+		let target = self.target().generate_syntax(context, &scope)?;
+
+		Ok(syntax::Alias {
+			ident: self.ident().clone(),
+			params,
+			target,
+		})
 	}
 }
